pyproject.toml
[project]
name = "neokline-private"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "seaborn==0.13.2",
    "mplfinance==0.12.10b0",
    "openpyxl==3.1.2",
    "xlrd==2.0.1",
    "st-annotated-text==4.0.1",
    "streamlit>=1.41.0",
    "streamlit-authenticator>=0.2.3",
    "streamlit-local-storage>=0.0.25",
    "streamlit-modal>=0.1.2",
    "tables==3.9.2",
    "tushare==1.4.21",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
    "python-docx==1.1.2",
    "pdfplumber==0.11.6",
    "faiss-cpu==1.10.0",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.ruff]
line-length = 120
exclude = [".venv", "data", "result", "target"]

[tool.ruff.lint]
extend-fixable = ["I"]
extend-select = ["I"]
extend-unfixable = ["F401"]

[tool.ruff.lint.isort]
length-sort = true
